Pushing working solution!
