+++ app/login/page.tsx
@@ -3,7 +3,7 @@ import { headers } from "next/headers";
 import { createClient } from "@/utils/supabase/server";
 import { track } from '@vercel/analytics/server';
 import { redirect } from "next/navigation";
-import { SubmitButton } from "./submit-button";
+import { SubmitButton } from "../forgot-password/submit-button";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 
@@ -12,14 +12,13 @@ export default function Login({
 }: {
   searchParams: { message: string };
 }) {
-  const defaultUrl = process.env.NEXT_PUBLIC_BASE_URL || "http://localhost:3000";
-
+  
   const signIn = async (formData: FormData) => {
     "use server";
-
+    
+    const supabase = createClient();
     const email = formData.get("email") as string;
     const password = formData.get("password") as string;
-    const supabase = createClient();
 
     const { error } = await supabase.auth.signInWithPassword({
       email,
@@ -37,10 +36,10 @@ export default function Login({
   const signUp = async (formData: FormData) => {
     "use server";
 
+    const supabase = createClient();
     const origin = headers().get("origin");
     const email = formData.get("email") as string;
     const password = formData.get("password") as string;
-    const supabase = createClient();
 
     const { error } = await supabase.auth.signUp({
       email,
@@ -55,22 +54,6 @@ export default function Login({
     }
 
     return redirect("/home");
-  };
-
-  const resetPassword = async (formData: FormData) => {
-    "use server";
-
-    const email = formData.get("reset_email") as string;
-    const supabase = createClient();
-
-    const { error } = await supabase.auth.resetPasswordForEmail(email, {
-      redirectTo: `${defaultUrl}/reset-password`,
-    });
-
-    if (error) {
-      return redirect("/login?message=Could not reset password");
-    }
-    return redirect("/login?message=Password reset email sent");
   };
 
   return (
@@ -108,27 +91,15 @@ export default function Login({
         >
           Sign Up
         </SubmitButton>
-        {searchParams?.message && (
-          <p className="mt-4 p-4 bg-foreground/10 text-foreground text-center">
-            {searchParams.message}
-          </p>
-        )}
-      </form>
 
-      <form className="flex-1 flex flex-col w-full justify-center gap-2 mt-6" action={resetPassword}>
-        <Label className="text-md" htmlFor="reset_email">
-          Forgot Password?
-        </Label>
-        <Input
-          className="rounded-md px-4 py-2 bg-inherit border mb-6"
-          name="reset_email"
-          placeholder="Enter your email"
-          required
-        />
-        <SubmitButton formAction={resetPassword} pendingText="Sending Reset Email...">
-          Reset Password
-        </SubmitButton>
+        <p className="mt-4 text-sm text-center">
+          Forgot your password?{" "}
+          <Link href="/forgot-password" className="text-blue-500 underline">
+            Reset Password
+          </Link>
+        </p>
       </form>
+      {searchParams?.message && <p className="text-center">{searchParams.message}</p>}
     </div>
   );
 }
