+++ app/reset-password/page.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { createClient } from "@/utils/supabase/server";
 import { redirect } from "next/navigation";
-import { SubmitButton } from "../login/submit-button";
+import { SubmitButton } from "../forgot-password/submit-button";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 
@@ -14,7 +14,6 @@ export default function Login({
     "use server";
 
     const code = searchParams.code;
-    const email = formData.get("email") as string;
     const password = formData.get("password") as string;
     const supabase = createClient();
 
@@ -22,7 +21,7 @@ export default function Login({
       return redirect("/forgot-password?message=Reset token or email is missing.");
     }
 
-    const { data, error: sessionError  } = await supabase.auth.exchangeCodeForSession(code);
+    const { error: sessionError  } = await supabase.auth.exchangeCodeForSession(code);
 
     if (sessionError) {
         console.error("Session error:", sessionError.message);
@@ -30,7 +29,6 @@ export default function Login({
       }
 
     const { error: updateError } = await supabase.auth.updateUser({
-        // email: email,
         password: password,   
     });
 
@@ -46,18 +44,7 @@ export default function Login({
     <div className="flex-1 flex flex-col w-full px-8 py-16 sm:max-w-md justify-center gap-2">
       <form
         className="flex-1 flex flex-col w-full justify-center gap-2 text-foreground"
-        
       >
-        {/* <Label className="text-md" htmlFor="email">
-          Email
-        </Label>
-        <Input
-          className="rounded-md px-4 py-2 bg-inherit border mb-6"
-          type="email"
-          name="email"
-          placeholder="you@example.com"
-          required
-        /> */}
         <Label className="text-md" htmlFor="password">
           New Password
         </Label>
